requirements_test.txt
atomicwrites==1.4.1
attrs==21.4.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-periphery==2.3.0
tomli==2.0.1
bitstring==3.1.9
pytest-mock==3.7.0
pytest-cov==3.0.0
protobuf==3.20.3